Move Song to top-level components
